Update Justifier tests and implement Settings class
Added more unit tests to JustifierTest.  Fixed width() method on TextBlock to account
for all contained lines being offset.  Implemented the nextName() method on Settings to
increment default names and made Settings a singleton.

Added document to top level of repo for tracking things that still need to get finished
for the conference.